refactor: rewrite code to minimize num of functions
