+++ assets/js/createBody.js
@@ -71,4 +71,15 @@ export default function createBody() {
   addNode('div', 'keyboard__info', document.querySelector('.wrapper.keyboard'));
   addNode('p', 'keyboard__info-p', document.querySelector('.keyboard__info'), 'Клавиатура создана в ОС Windows.');
   addNode('p', 'keyboard__info-p', document.querySelector('.keyboard__info'), 'Переключение раскладки: Left Ctrl + Left Alt.');
+
+  const idsRow1 = ['Backquote', 'Digit1', 'Digit2', 'Digit3', 'Digit4', 'Digit5', 'Digit6', 'Digit7', 'Digit8', 'Digit9', 'Digit0', 'Minus', 'Equal', 'Backspace'];
+  const idsRow2 = ['Tab', 'KeyQ', 'KeyW', 'KeyE', 'KeyR', 'KeyT', 'KeyY', 'KeyU', 'KeyI', 'KeyO', 'KeyP', 'BracketLeft', 'BracketRight', 'Backslash', 'Delete'];
+  const idsRow3 = ['CapsLock', 'KeyA', 'KeyS', 'KeyD', 'KeyF', 'KeyG', 'KeyH', 'KeyJ', 'KeyK', 'KeyL', 'Semicolon', 'Quote', 'Enter'];
+  const idsRow4 = ['ShiftLeft', 'IntlBackslash', 'KeyZ', 'KeyX', 'KeyC', 'KeyV', 'KeyB', 'KeyN', 'KeyM', 'Comma', 'Period', 'Slash', 'ArrowUp', 'ShiftRight'];
+  const idsRow5 = ['ControlLeft', 'MetaLeft', 'AltLeft', 'Space', 'AltRight', 'ControlRight', 'ArrowLeft', 'ArrowDown', 'ArrowRight'];
+  const ids = idsRow1.concat(idsRow2, idsRow3, idsRow4, idsRow5);
+  const btns = document.querySelectorAll('.keyboard__btn');
+  for (let i = 0; i < btns.length; i += 1) {
+    btns[i].dataset.code = ids[i];
+  }
 }
